refactor(login-guard): adapted login guard and added 7 day login check #31
